エージェントを複数起動しないように修正
(1) 既に起動している場合はそれを再利用する

+++ run-agent.sh
@@ -2,10 +2,25 @@
 
 #set  -xu
 
-ssh-agent | tee  .agent
-eval  `cat  .agent`
+# ディレクトリを確実に存在させて、パーミッションを設定する
+agent_lnk='.agent'
+agent_dir=$(dirname $(readlink -f ${agent_lnk}))
+if [[ ! -d ${agent_dir} ]] ; then
+    mkdir -p ${agent_dir}
+fi
+chmod 700 ${agent_dir}
+
+flag_add=FALSE
+if [[ ! -f ${agent_lnk} ]] ; then
+    ssh-agent | tee .agent
+    flag_add=TRUE
+fi
+chmod 600 ${agent_lnk}
+
+eval  $(cat  ${agent_lnk})
 
 #set  +xu
 
-ssh-add  ~/.ssh/*.ecdsa
-
+if [[ "x${flag_add}y" = 'xTRUEy' ]] ; then
+    ssh-add  ~/.ssh/*.ecdsa
+fi
